Add Serializable DTO classes for the v1 HTTP API

+++ lightweaver-core/src/main/kotlin/com/oelderoth/lightweaver/core/colorsource/HsvMeanderColorSource.kt
@@ -4,6 +4,6 @@ import com.oelderoth.lightweaver.core.colors.Color
 import com.oelderoth.lightweaver.core.pixeloffsets.PixelOffset
 
 class HsvMeanderColorSource(uid: Int, val meander: HsvMeanderConfig, override val pixelOffsets: PixelOffset) : ColorSource(uid), AddressableColorSource {
-    class MeanderRange(val duration: Int, val distance: Int)
+    class MeanderRange(val duration: Int, val distance: Float)
     class HsvMeanderConfig(val color: Color, val hue: MeanderRange, val saturation: MeanderRange, val value: MeanderRange)
 }
